+++ main.py
@@ -1,10 +1,10 @@
-# from interface.cli.main import app
-from interface.ui.login_page import start
+from interface.cli.app import start as cli
+from interface.ui.login_page import start as gui
 
 
 def main():
-    # app()
-    start()
+    gui()
+    # cli()
 
 
 main()